Add Cypress with default files
Adding Cypress to be able to test the projects. I am currently keeping
the default files since I need to learn how it works at all.
Linting for Cypress is currenty ignored. This is because of the
overwhelming amount of errors to resolve. Should be eaiser to fix once
all default files are removed.