Add user panel to header
Inside the user panel I've added settings for courtage and currency
and updates those values in localStorage. However when updating
localStorage the home-component doesn't rerender on updates in
localStorage (as it should), so to create a better flow redux will
be added in a separate commit to complete the functionality.